app/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Django==4.0.4
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
huggingface-hub==0.5.1
idna==3.3
joblib==1.1.0
numpy==1.22.3
packaging==21.3
pyparsing==3.0.8
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
six==1.16.0
sqlparse==0.4.2
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.18.0
typing-extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9